feat: undo creating data
